+++ include/aikido/planner/World.hpp
@@ -40,7 +40,7 @@ public:
 
   virtual ~World();
 
-  /// Create a new World inside of a shared_ptr
+  /// Create a new World ptr
   /// \param name Name for the new World
   static std::unique_ptr<World> create(const std::string& name = "");