+++ src/statespace/dart/MetaSkeletonStateSpace.cpp
@@ -283,6 +283,24 @@ void MetaSkeletonStateSpace::checkCompatibility(
 }
 
 //==============================================================================
+void MetaSkeletonStateSpace::checkIfContained(
+    const ::dart::dynamics::Skeleton* skeleton) const
+{
+  // TODO: Name-uniqueness is allowed only within the same skeleton,
+  // so we should check for skeleton-equality.
+  auto dofNames = mProperties.getDofNames();
+  for (const auto& name : dofNames)
+  {
+    if (!skeleton->getDof(name))
+    {
+      std::stringstream ss;
+      ss << "DegreeOfFreedom[" << name << "] does not exist in mSkeleton.";
+      throw std::invalid_argument(ss.str());
+    }
+  }
+}
+
+//==============================================================================
 void MetaSkeletonStateSpace::convertPositionsToState(
     const Eigen::VectorXd& _positions, State* _state) const
 {
@@ -344,11 +362,49 @@ auto MetaSkeletonStateSpace::getScopedStateFromMetaSkeleton(
 
 //==============================================================================
 void MetaSkeletonStateSpace::setState(
-    ::dart::dynamics::MetaSkeleton* _metaskeleton, const State* _state)
+    ::dart::dynamics::MetaSkeleton* _metaskeleton, const State* _state) const
 {
   Eigen::VectorXd positions;
   convertStateToPositions(_state, positions);
   _metaskeleton->setPositions(positions);
+}
+
+//==============================================================================
+::dart::dynamics::MetaSkeletonPtr
+MetaSkeletonStateSpace::getControlledMetaSkeleton(
+    const ::dart::dynamics::SkeletonPtr& skeleton) const
+{
+  using ::dart::dynamics::Group;
+  using ::dart::dynamics::DegreeOfFreedom;
+
+  std::vector<DegreeOfFreedom*> dofs;
+  dofs.reserve(mProperties.getNumDofs());
+
+  for (const auto& dofName : mProperties.getDofNames())
+  {
+    DegreeOfFreedom* dof = skeleton->getDof(dofName);
+    if (!dof)
+    {
+      throw std::invalid_argument(
+          "Skeleton has no DegreeOfFreedom named '" + dofName + "'.");
+    }
+    dofs.emplace_back(dof);
+  }
+
+  auto controlledMetaSkeleton
+      = Group::create(mProperties.getName(), dofs, false, true);
+  if (!controlledMetaSkeleton)
+  {
+    throw std::runtime_error(
+        "Failed creating MetaSkeleton of controlled DOFs.");
+  }
+
+  if (controlledMetaSkeleton->getNumDofs() != mProperties.getNumDofs())
+  {
+    throw std::runtime_error("Only single-DOF joints are supported.");
+  }
+
+  return controlledMetaSkeleton;
 }
 
 } // namespace dart